worker/requirements.txt
numpy
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
openai==1.6.1
fastapi==0.109.0
SQLAlchemy
asyncpg==0.29.0
pydantic==2.7.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
redis==5.0.1
celery==5.3.6
tldextract==5.1.1
httpx==0.26.0
filelock==3.13.1
faiss-cpu==1.7.4